feat(mesh): add face set ops and delimit support
- Add BMO_DELIM_FACE_SET delimit flag for mesh selection operations
- Implement four new face set visibility operators: hide active face set, hide inactive face sets, show all face sets, isolate face set
- Add UI menu entries for the new operators in the 3D View's show/hide menu
- Update linked face selection logic to respect face set delimiter

+++ source/blender/draw/intern/draw_cache_extract.hh
@@ -231,7 +231,7 @@ enum DRWBatchFlag : uint64_t {
   MBC_WIRE_LOOPS_UVS = (1u << MBC_BATCH_INDEX(wire_loops_uvs)),
   MBC_WIRE_LOOPS_EDITUVS = (1u << MBC_BATCH_INDEX(wire_loops_edituvs)),
   MBC_SCULPT_OVERLAYS = (1u << MBC_BATCH_INDEX(sculpt_overlays)),
-  MBC_VIEWER_ATTRIBUTE_OVERLAY = (1u << MBC_BATCH_INDEX(surface_viewer_attribute)),
+  MBC_VIEWER_ATTRIBUTE_OVERLAY = (uint64_t(1u) << MBC_BATCH_INDEX(surface_viewer_attribute)),
   MBC_PAINT_OVERLAY_VERTS = (uint64_t(1u) << MBC_BATCH_INDEX(paint_overlay_verts)),
   MBC_PAINT_OVERLAY_SURFACE = (uint64_t(1u) << MBC_BATCH_INDEX(paint_overlay_surface)),
   MBC_SURFACE_PER_MAT = (uint64_t(1u) << MBC_BATCH_LEN),
